make `find_signature` generic
