+++ blescan-discovery/src/local.rs
@@ -3,11 +3,11 @@ use std::error::Error;
 use std::time::Duration;
 use tokio::time;
 
-use btleplug::api::{Central, Manager as _, Peripheral, PeripheralProperties, ScanFilter};
+use btleplug::api::{Central, Manager as _, Peripheral as BtlePeripheral, ScanFilter};
 use btleplug::platform::{Adapter, Manager};
 
 use blescan_domain::discover::DiscoveryEvent;
-use blescan_domain::signature::Signature;
+use blescan_domain::peripheral::{find_signature, Peripheral};
 
 use crate::Scanner;
 use async_trait::async_trait;
@@ -41,7 +41,11 @@ impl Scanner for LocalScanner {
         let current_time = Utc::now();
         for peripheral in &peripherals {
             let properties = peripheral.properties().await?.unwrap();
-            if let Some(signature) = find_signature(&properties)
+            let peripheral_info = Peripheral::new(
+                properties.local_name.clone(),
+                properties.manufacturer_data.clone(),
+            );
+            if let Some(signature) = find_signature(&peripheral_info)
                 && let Some(rssi) = properties.rssi
             {
                 events.push(DiscoveryEvent::new(current_time, signature, rssi));
@@ -49,23 +53,5 @@ impl Scanner for LocalScanner {
         }
         self.adapter.stop_scan().await.expect("Can't stop scan");
         Ok(events)
-    }
-}
-
-pub fn find_signature(properties: &PeripheralProperties) -> Option<Signature> {
-    if let Some(local_name) = &properties.local_name {
-        Some(Signature::Named(local_name.clone()))
-    } else if !&properties.manufacturer_data.is_empty() {
-        let mut context = md5::Context::new();
-        let mut manufacturer_ids: Vec<&u16> = properties.manufacturer_data.keys().collect();
-        manufacturer_ids.sort();
-        for manufacturer_id in manufacturer_ids {
-            let arbitrary_data = properties.manufacturer_data[manufacturer_id].clone();
-            context.consume(arbitrary_data);
-        }
-        let digest = context.compute();
-        Some(Signature::Anonymous(format!("{digest:x}")))
-    } else {
-        None
     }
 }
