+++ blescan-domain/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod chrono_extra;
 pub mod device_state;
 pub mod discover;
+pub mod peripheral;
 pub mod signature;
 pub mod snapshot;
 pub mod state;
